author relatrd posts

+++ admin/includes/view_all_comments.php
@@ -70,7 +70,7 @@ $result=mysqli_query($connection,$query);
         echo "<td><a href='comments.php?approve=$comment_id'>Approve</a> </td>";
     echo "<td><a href='comments.php?unapprove=$comment_id'>Unapprove</a> </td>";
     
-    echo "<td><a href='comments.php?delete=$comment_id'>Delete</a> </td>";
+    echo "<td><a onClick=\"javascript: return confirm('Are you sure you want to delete this comment'); href='comments.php?delete=$comment_id'>Delete</a> </td>";
         echo "</tr>";
     //echo "<td><a href='categories.php?delete=$cat_id'>Delete</a></td>";
     //echo "<td><a href='categories.php?edit=$cat_id'>Edit</a></td></tr>";
