+++ admin/includes/view_all_posts.php
@@ -11,8 +11,32 @@ if(isset($_POST['checkBoxArray'])){
                  $query="UPDATE posts SET post_status='{$bulk_options}' WHERE post_id = {$postValueID}";
                 break;
                 // case'draft':
+                 case'clone':
+                        $query="SELECT * FROM posts WHERE post_id= {$postValueID}";
+                        $select_post_query=mysqli_query($connection,$query);
+                        confirmQuery($select_post_query);
+                        while($row=mysqli_fetch_assoc($select_post_query)){
+                                $post_title=$row['post_title'];
+
+                                $post_category_id=$row['post_category_id'];
+                                $post_author=$row['post_author'];
+                                $post_date=$row['post_date'];
+                                $post_image=$row['post_image'];
+                                //$post_comments_count=$row['post_comments_count'];
+                                $post_status=$row['post_status'];
+                                $post_tags=$row['post_tags'];
+                                $post_content=$row['post_content'];
+
+                        }
+                        $query="INSERT INTO posts (post_title,post_category_id,post_author,post_date,post_image,post_status,post_tags,post_content) ";
+                        $query.=" VALUES ('{$post_title}',{$post_category_id},'{$post_author}',now(),'{$post_image}','{$post_status}','{$post_tags}', '{$post_content}')";
+//                         $clone_post_query=mysqli_query($connection,$query);
+//                       if(!$clone_post_query){
+//                           die("clone query failed ".mysqli_error($connection));
+//                       }
                 
-                //break;
+                
+                 break;
                  case'delete':
                 $query="DELETE FROM posts WHERE post_id = {$postValueID} ";
                 break;
@@ -30,6 +54,7 @@ if(isset($_POST['checkBoxArray'])){
                                    <option value="">Select Options</option>
                                    <option value="published">Publish</option>
                                    <option value="draft">Draft</option>
+                                   <option value="clone">Clone</option>
                                    <option value="delete">Delete</option>
                                </select>
                            </div>
@@ -49,8 +74,10 @@ if(isset($_POST['checkBoxArray'])){
                                     <th>Image</th>
                                     <th>Tags</th>
                                     <th>Comments</th>
+                                    <th>Views</th>
                                     <th>Date</th>
                                     <th>content</th>
+                                    <th>&nbsp;</th>
                                     <th>&nbsp;</th>
                                     <th>&nbsp;</th>
                                     <th>&nbsp;</th>
@@ -59,7 +86,7 @@ if(isset($_POST['checkBoxArray'])){
                             <tbody>
                                
 <?php
- $query="SELECT * FROM posts";
+ $query="SELECT * FROM posts ORDER BY post_id DESC";
     $select_all_from_posts=mysqli_query($connection,$query);
 
     while($row=mysqli_fetch_assoc($select_all_from_posts)){
@@ -73,6 +100,7 @@ if(isset($_POST['checkBoxArray'])){
     $post_status=$row['post_status'];
     $post_tags=$row['post_tags'];
     $post_content=$row['post_content'];
+    $post_views_count=$row['post_views_count'];
         
         echo "<tr>";
         ?>
@@ -96,11 +124,13 @@ if(isset($_POST['checkBoxArray'])){
     echo "<td><img width='100px' src='../images/$post_image'/></td>";
     echo "<td>$post_tags</td>";
     echo "<td>$post_comments_count</td>";
+    echo "<td> <a href='posts.php?reset=$post_id'>$post_views_count</a></td>";
     echo "<td>$post_date</td>";
     echo "<td>$post_content</td>";
     echo "<td><a href='posts.php?source=edit_post&p_id=$post_id'>Edit</a> </td>";
-    echo "<td><a href='posts.php?delete=$post_id'>Delete</a> </td>";
+    echo "<td><a onClick=\"javascript: return confirm('Are you sure you want to delete this post'); \" href='posts.php?delete=$post_id'>Delete</a> </td>";
     echo "<td><a href='../post.php?p_id=$post_id'>View Post</a> </td>";
+   // echo "<td><a href='posts.php?reset=$post_id'>Reset Views</a> </td>";
         echo "</tr>";
     //echo "<td><a href='categories.php?delete=$cat_id'>Delete</a></td>";
     //echo "<td><a href='categories.php?edit=$cat_id'>Edit</a></td></tr>";
@@ -125,11 +155,21 @@ if(isset($_POST['checkBoxArray'])){
 
 if(isset($_GET['delete'])){
     
-    $get_post_id=$_GET['delete'];
+    $get_post_id=mysqli_real_escape_string($connection,$_GET['delete']);
     
     $query=" DELETE FROM posts WHERE post_id = {$get_post_id}";
     $delete_query_result=mysqli_query($connection,$query);
     confirmQuery($delete_query_result);
+    header("Location: posts.php");
+}
+
+if(isset($_GET['reset'])){
+    
+    $get_post_id=mysqli_real_escape_string($connection,$_GET['reset']);
+    
+    $query=" UPDATE posts  SET post_views_count =0 WHERE post_id = {$get_post_id}";
+    $update_query_result=mysqli_query($connection,$query);
+    confirmQuery($update_query_result);
     header("Location: posts.php");
 }
