+++ admin/includes/view_all_users.php
@@ -78,7 +78,7 @@ $result=mysqli_query($connection,$query);
     echo "<td><a href='comments.php?delete=$comment_id'>Delete</a> </td>";*/
        echo "<td><a href='users.php?change_to_admin=$user_id'>Admin</a></td>";
     echo "<td><a href='users.php?change_to_subscriber=$user_id'>Subscriber</a></td>";
-    echo "<td><a href='users.php?delete=$user_id'>Delete</a></td>";
+    echo "<td><a onClick=\"javascript: return confirm('Are you sure you want to delete this user'); href='users.php?delete=$user_id'>Delete</a></td>";
     echo "<td><a href='users.php?source=edit_user&u_id=$user_id'>Edit</a></td>";
     //echo "<td><a href='categories.php?edit=$cat_id'>Edit</a></td></tr>";
         echo "</tr>";
